Update safe areas without going through itemTransformChanged if possible
When the QWindow's safe area margins change, or additional margins are
set on a QQuickSafeArea, we need to propagate this to any affected
safe areas. Doing it via itemTransformChanged worked, because we were
already hooking into that to detect when an item is moved, but was
making it hard to reason about whether the transform change could be
deferred until the corresponding geometry change.

We now use an explicit function to update safe areas, which also
works in paint order, as opposed to itemTransformChanged which
delivers in reverse paint order (a bug, that will hopefully be fixed).

Pick-to: 6.9
Change-Id: If2e033dfa0306ff5980b43e9a1e5721d7ec0e3da
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>