QtQml: Simplify type loader thread life cycle
We don't have to keep the thread object around at all while the thread
is not running. Just deleting and re-creating it tremendously simplifies
things. Furthermore, we can now create it on demand the first time it's
needed. This way we gain the ability to determine whether it's still
safe to initialize values the thread will later read without locking.

Task-number: QTBUG-131721
Change-Id: I702c563ac744387e1be82364b3e87874741d89f4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>