Baseline tests: move mouse cursor out of the way
Some controls render in a hovered state when they are under the mouse
pointer, resulting in flakiness if we don't (try to) control where the
mouse is when the window shows.

Move it to the top left of the screen's available geometry which will
typically be over the titlebar of the test window.

Change-Id: I949250b884dac5294e5034513074097b3133084e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>