Software Renderer: Ensure flushing of previously dirty regions
Previously, if an opaque item was moved completely behind another opaque
item between frames, the previous region occupied by the item might not
be flushed. This occurred because, when subtracting the obscured region
from the node’s dirty region, the dirty region could become empty. In
such cases, the node would mark itself as no longer dirty, resulting in
no flushing for the next frame. This could cause the old position of the
item not to be cleared properly.

To address this issue, additional logic has been added to ensure that
the previous dirty region is still flushed, even in this edge case.

Fixes: QTBUG-132192
Pick-to: 6.9 6.8 6.5
Change-Id: Iad45ddeeb65e70090e8368d94245b13692017fd1
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>