QmlDom. Remove loadPendingDependencies call from DomItem::fromCode
This commit is a minor simplification of DomItem::fromCode to remove
a redundant and misleading call to loadPendingDependencies.
When Env is created with NoDependencies option, the load of the File is
happening instantly, not through the loading queue, therefore there are
no pending dependencies.

+ Minor cosmetic changes.
Fore example, removing Option::SingleThreaded option, it's not used anywhere, like it's actually redundant and we should remove it completely. But it's also misleading. In most cases we create DomEnv with SingleThreaded option, but then we actually using lots of locks inside environment and it seems that inside Qmlls we indeed have more than 1 thread.

The similar was happening to DomUniverse from which I've successfully removed this option

Change-Id: Ia8e9707fd70dcc319330911ad9ea00df86629636
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>