QtQml: Compare before assigning to var properties
We don't need to assign anything if the values are strictly equal in the
JavaScript sense. In that case they are indistinguishable and no signal
should be sent either.

[ChangeLog][QtQml][Important Behavior Changes] Assigning to "var"
properties now only triggers a change signal if the value of the
property actually changes. Change is determined using JavaScript's
"strictly equals" semantics (the "===" operator).

Fixes: QTBUG-131901
Change-Id: I42b8f370ea7e10d684aa279db9bff33d78253516
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>