Focus frame: Don't delete focus frame whenever new focus item is null
When we didn't find a new target focus item we would destroy the focus
frame object and recreate it the next time a new valid focus item is
found. This can be expensive due to cleanup and reallocation calls,
and it can also lead to a crash during parent destruction because the
focus frame gets deleted as part of the parent item's destruction,
potentially triggering changes in the focus object that result in
another attempted deletion of the focus frame.
Instead, in case of an invalid new focus item, we should just set the
focus frame's target item to null when calling moveToItem. This
will handle that case by making the frame invisible and setting its
parent to null.

Pick-to: 6.9 6.8
Change-Id: Icf46af0ce3c79bab09343a2262688fb33ed8a4d9
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>