Remove safe area change listener on all items on destruction
We listen to transform changes on the item we're attached to, but
also to geometry changes on any item that affects the transform
of the attached item. During destruction of the safe area the
geometry change items may still be alive, which would end up
with a crash if they try to notify the dead safe area of changes.

We now track the items we listen to, and remove ourselves before
the safe area is destroyed.

Pick-to: 6.9
Change-Id: I9cc5197b9dabcbb40a23c6094c7989fe480b349d
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>