FocusFrame: Don't remove binding on visible property
When calling moveToItem() when the focus object changes, we were
explicitly assigning the visible property. This caused the previous
binding declared on the visible property of the frame to be removed.
To fix, don't assign the visible property on moveToItem, as the
binding on the target item and its visibility will take care of that.
Ammends 945c3a98edac75cac082b1719c380a0bae861944.

Pick-to: 6.8 6.9 6.5
Change-Id: I504a0e00000c5fdfb8c5193b9fa2314199a7c4e0
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>