Fix verification of jsroot metatypes
This is in fact a test for qmltyperegistrar and it makes no sense to
check the qmltypes. Those can only be generated by qmltyperegistrar's
--jsroot argument which will use the built-in metatypes. The interesting
part is whether the metatypes are correct. Check that instead.

Amends commit 578946437fb31edd6aa78612e9db8fed9a3eb210 which introduced
an attribute to qmltypes that can only be generated with the --jsroot
argument to qmltyperegistrar.

Pick-to: 6.9 6.8
Change-Id: Iae9f7d08980ecb09f61000b7f63d6d7e9a746970
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>