CMake: Fix find_package call in Qt6QmlFindQmlscInternal.cmake
Qt6QmlFindQmlscInternal.cmake tries to find the
Qt6QmlCompilerPlusPrivateTools package and specifies a version. That
version was PROJECT_VERSION, which is the version of the user project.
It should be the Qt's version instead.

Read the package version of the Qt6::Qml target and use that in the
find_package call. The target is guaranteed to exist, because the
Qt6QmlFindQmlscInternal.cmake inclusion is guarded by a check for this
target.

This amends commit b0f1ec4e394dcee82400964225be485a7cdd3c53.

Pick-to: 6.8 6.9
Fixes: QTBUG-132421
Change-Id: I04402296b351ef8df991ade697c633594b0b17b5
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>