QQmlBind: Split object/property/value and generalized group modes
There is not much overlap in either the functionality or the data they
need. We will need to introduce a third mode, so this is a good time to
clean up.

Also, finally add change signals to all the properties. Omitting the
change signals from publicly accessible properties is rather rude. Add
const to the object() accessor.

Task-number: QTBUG-132541
Change-Id: Ibba362e84a8bedf1322a5d2fc05010f5e1fe481f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>