Fix flakiness in tst_qquicktextedit
On platforms running X11, windows are activated in an asynchronous
fashion. It is necessary to wait for the activation to happen using
primitives in the Qt Test library. This eliminates possible flakiness.
Such flakiness is seen on Ubuntu and OpenSuse.

Use QTRY_VERIFY instead of QVERIFY. This means that we wait five seconds
for window activation to happen.

Task-number: QTBUG-130374
Change-Id: I529b8d868bd3ecb5874ea8e9b950884ce1f8074e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>