Set window position to fix tst_TapHandler::longPress flakiness
tst_QQuickTaphandler::longPress() was flaky due to a qquickview with a
non-fixed position. When failing the mouse event was not received by
the button.

The change in position of the qquickview caused the mouse event to
sometimes be done higher than it should. The frame position was given
instead of the position of the widget itself, causing the issue.

Give a fix position and verify it with QTRY_COMPARE to ensure that the
position of the widget is correct and that the mouse event will be
delivered on the button.

Remove tst_QQuickTaphandler::longPress() from blacklisted test as it
is not flaky anymore.

Fix flakiness on Opensuse

Fixes: QTBUG-132628
Change-Id: Ie1aba47e16377c2e8b3bab109d356708797001cc
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>