qmlpreview: Do not create file loader until needed
... and delete it when not needed anymore.

While the preview service is inactive we don't need the file loader.
Since the file loader is rather heavyweight and comes with its own
thread, we should limit its life time.

Furthermore, it is suspected that the termination of the file loader
thread in the file loader's dtor triggers a bug that causes the debugger
tests to hang. Terminating the thread already on state change might work
around it.

Pick-to: 6.9 6.8
Task-number: QTBUG-101678
Task-number: QTBUG-101972
Task-number: QTBUG-102984
Task-number: QTBUG-132697
Change-Id: I244aa4c3d4257f0c4b8f46f826710a76cf389baa
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>