qmldom: make Path methods less deceitful
Stop the confusion between Path::field() and Path::Field() and just
rename
* Path::Field() to Path::fromField() and
* Path::field() to Path::withField()

Previously, Path::field() would return a new Path with the Field
appended to to the current path, while Path::Field() would create a new
Path only containing the passed Field, which I think is a bit confusing.

Same for the other methods returning copies with components appended
to it, and their uppercased counterparts:
* Path::empty()
* Path::key()
* Path::index()
* Path::any()
* Path::filter()
* Path::current()
* Path::path()

I won't explain here why I believe how Path::empty(), for example, was
particularly deceitful with its previous name.

Also rename appendComponent to withComponent that actually does not
append anything: it returns a copy with the component appended to it.

Change-Id: I9120c3827e30330f715e699cbbb88780d2fd358d
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>