FluentWinUI3: Move FocusFrame type to private impl module
FocusFrame.qml was included in the FluentWinUI3 public module imports,
despite not providing any documentation. This component is an
implementation detail and should not be part of public QML API.
Add deprecation warnings in the original file and move it to the
private impl module instead.

[ChangeLog][QtQuick][Controls][FluentWinUI3] FocusFrame.qml is deprecated
in the module's public QML API and moved to the private impl module instead.

Pick-to: 6.9
Change-Id: I74c8c9a03ccc1f7b33d6fb814ca1faeca57b0545
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>