Isolate Flickable's content item from Flickable's safe area margins
Items in a Flickable's content item that account for safe area margins
will behave unexpectedly when the flickable is scrolled, so we isolate
the flickable's content item from the Flickable itself and any safe
area margins that are inherited from its parent items or window.

This matches the behavior of UIScrollView when the content inset
adjustment behavior is anything but "none". We should consider a
similar granular approach if/when we add logic to adopt the safe
area margins into the ScrollArea's content margins for Controls.

Fixes: QTBUG-132737
Pick-to: 6.9
Change-Id: I13cfb3b03cccc6e9aedb645e0e370e4a36c20423
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>