Deliver QContextMenuEvent via ExtraData.contextMenu, not event filter
Event filtering can be problematic. At least this was a more reasonable
use case for a filter than some others in Controls; still, we already
have a switch case in QQuickItem::event(), and plan to add a virtual
QQuickItem::contextMenuEvent() in Qt 7, so we might as well mock it up
now as QQuickItemPrivate::contextMenuEvent(). Dispatching only the
QContextMenuEvents ought to be cheaper than filtering all events.
On the downside, ExtraData gets a little bigger.

We add QQuickItemPrivate::setContextMenu() because
QQuickItem::ExtraData is not exported; and to enable categorized
logging when one menu is replaced with another, this setter returns
the previously-known menu, if any.

Pick-to: 6.9
Change-Id: I9f2553fb579409becf797046dcc473260320c6a5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>