qmlls: librarify into QQmlLSPrivate
Move the tool's code to the library and make qmllanguageservertool.cpp
as small as possible. Fix the strings to have _L1 to avoid warnings.

This would allow to have a duplicate version of
qmllanguageservertool.cpp outside of qtdeclarative, which then could be
statically linked to the help plugin in qttools. This would enable the
documentation support in the standalone qmlls build (which currently
can't provide any documentation hints).

Task-number: QTBUG-132773
Change-Id: Ia1717c0f71f70f6bb5d863017fe050117688d9e7
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>