Process QQuickItemPrivate::transformChanged before recursing children
The child recursion logic in transformChanged notes that we "Inform the
children in paint order: by the time we visit leaf items, they can see
any consequences in their parents".

This holds true for subclasses of QQuickItemPrivate, that process the
call and then call the base class implementation.

However in the case of updating the layer matrix we were doing it after
recursing children, which breaks the documented logic.

Similarly, we were notifying QQuickItem::ItemTransformHasChanged
after recursion, which meant that observers would get callbacks in
reverse paint order when observing a hierarchy of items.

We now handle both these cases before recursing.

Pick-to: 6.9
Change-Id: I373a2d7af11da65f66494322e44079ae04e95752
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>