Ensure safe areas are always updated in paint order
We can't rely on the transform and geometry change listener callbacks
to come in in paint order, but we want the safe areas to update in
paint order, as the margins of a parent safe area may affect children.

To work around the order of the transform and geometry callbacks we
find the safe area closest to the item that was transformed, and let
that safe area do a recursive update.

Pick-to: 6.9
Change-Id: I1221166ba6ccafb5fed78370606e4ccd1c26cbd9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>