QmlCompiler: Reduce duplication of "not found" messages
Track the unresolved types in a central place and only warn once for
each of them. Make sure that we still mark them as wrapped in implicit
components even if we don't warn, though.

Realize that custom parser parents potentially obfuscate any types
defined inside their scope, not only the ones we look for when
setting bindings.

Pick-to: 6.9 6.8
Task-number: QTBUG-124913
Change-Id: I30d911dc92d1e4359db66671bf62393f2f852b6c
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>