QtQml: Ignore revisions on signals of attached objects
If you attach a handler to such a signal, the operation is qualified.
There is no need to observe revisions. Furthermore, the context revision
to be observed there was a substantial source of confusion.

Amends commit 30db7d94e002b5c5b29820050a2220ef06afa54c which was
mistaken regarding availability of grouped properties.

Pick-to: 6.9
Fixes: QTBUG-65463
Change-Id: I3e695c0a8f0cbb779230e80889b83b7d7dfbf29b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>