Make SafeArea.additionalMargins actually add to the margins, not union
The semantics of SafeArea.additionalMargins was initially to union the
additional margins with the inherited margins, as this fit nicely with
the case of adding margins for a header that lives in the non-safe area
of the window.

But for the case of a header that lives within the existing safe area,
but wants to add additional margins for its container's children, the
API was awkward, requiring the user to include the inherited margins
explicitly.

The name additionalMargins was also not correct given these semantics.

We now follow the semantics of addition, which makes the use-case of
a header that lives in the non-safe area a bit more explicit, but also
clearer on exactly how much additional margins is added.

The new semantics also match how AppKit/UIKit handles additional safe
area insets, so the behavior should be familiar to anyone coming from
those platforms.

Pick-to: 6.9
Change-Id: I05e42e68da4237c0c7963457452ffea034ea3ae1
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>