Verify window active in flaky test tst_qquicklistview2
tst_qquicklistview2::isCurrentItem_NoRegressionWithDelegateModelGroups
is flaky where in some cases the key press to the left is not received
by the window, making the test fail.

This is caused by the keyPress event being sent before the window has
become active.

Verify that the window is active before sending the keyPress event to
it.

Fix flakiness on Opensuse

Change-Id: I03b597379aa72db5d8f740abe0d9e78012dfd971
Reviewed-by: Magdalena Stojek <magdalena.stojek@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>