QQuickPositioner: re-engineer PositionedItem
The class violated the rule of 5, had insufficient encapsulation, and
triggered undefined behavior.

It was designed to be stored in a QPODVector, but when using
transitions, it owns a pointer (transitionableItem). However the class
didn't disengage copies (!), and was actively being copied around in
code (thankfully, holding a null pointer, otherwise this would
double-delete).

Since QPODVector doesn't run destructors when deleting elements
("POD", but PositionedItem isn't POD, and thus not running the
destructor is language UB), another couple of methods were needed when
removing elements from the vector, to make sure to delete the pointer.

Make PositionedItem hold a unique_ptr instead, for clear ownership
semantics. Since QPODVector won't work any more, store it in a
std::vector. There is no longer need of external code to manage
ownership, so remove it.

The rest of the changes are mostly mechanical: turn index-based
iterations into range-based for loops, and calls to index-based
QPODVector::find to iterator-based std::find. To ease the search,
add an overload of operator== that takes a QQuickItem (with similar
semantics to the existing operator==).

Change-Id: I5eed1a9aada4db2417a586f2113c6892c058fbb0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>