QPODVector: code tidies
The usefulness of this class is questionable, but in any case let's not
repeat the PositionedItem fiasco: assert that the type held in a
QPODVector is trivial (ly copiable, destructible, ...).

While at it, use the proper macro to disable its copies. (The class is
in principle copiable, but that's pre-existing).

Change-Id: I03017290193b4bde457a72644a449b514b6dd0e0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>