QtQml: Reset top-level-required flag after processing sub-objects
For objects assigned to proeprties we use a QQmlObjectCreator with the
same shared state as the top level QQmlObjectCreator. However, that
one's top level required properties are obviously not the ones of the
outer object creator.

This exposes a different bug: If objects assigned to properties of
delegates cannot be created due to unset required properties, we should
delete the delegate right away before adding it. Currently the relevant
code warns about the unset required properties but still leaves the
object intact.

Pick-to: 6.9 6.8
Fixes: QTBUG-133052
Task-number: QTBUG-133129
Change-Id: I208fcbffa17443ad62c4aaa81aaec291c8cdcc82
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>