QQmlJSTranslationFunctionMismatchCheck: improve local arrays
Use CTAD to deduce (in particular) the size of the arrays. No-one
should need to pass the size of an array in this day and age if the
compiler can do it for us. We had a case once where this had gone
wrong and resulted in UB because the array was zero-extended and broke
the is_sorted() precondition of a lower_bound() use on it. Here, we
could get False Positiives if the arrays contained empty entries.

Also port from QStringView to QLatin1StringView. This is more compact
and should (once we fix it QtCore) result in faster comparison than
even UTF-16/UTf-16.

Finally, don't make the arrays static, which forces the compiler to
allocate storage for them in the executable and may lead to
unnecessary relocations, but mark them as constexpr. This ensures that
no runtime initialization (incl. magic static thread protection) takes
place while giving the compiler leeway as to how to lay the thing out
in memory optimally.

Amends 8b61addfa42525c1bd5f27febbe62dcbdb645fc9.

Change-Id: I8949c52525569d11f9e00f47548a019b3ce73cd6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>