Spreadsheets: Prevent data loss when dropping cells in invalid locations
Before this patch, dropping selected cells into an invalid place in the
TableView could result in partial data loss. This occurred because some
of the selected data couldn't be set to the new cells, and the data was
copied at the start of the drag operation.

This patch improves the drag-and-drop functionality by:
- Checking the possibility of copying selected data to target cells
before the drop occurs (and during dragging).
- Avoiding highlighting of invalid target cells during the drag
operation.

Task-number: QTBUG-125767
Change-Id: Iff1058c2239f4bad3430f4f01ccea619f1ac8773
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>