Controls: improve ComboBox mouse tests
The ComboBox::test_mouse was too unconstrained, giving false positives.
The spirit of the test is to click on the second item and asserts that
it was activated. It did not check the resulting currentIndex value.
The FluentWinUI3 popup enter animation is animated, this caused the
test to instead click on the first item but still pass the test.
The test now waits for the enter animation to finish and checks the
currentIndex after mouse release.
This also adds another test making sure the activated signal does get
emitted when the current item is clicked in the popup.

This is a preliminary work for the patch making currentValue writable.

Task-number: QTBUG-100345
Change-Id: Ib75435552a65ad6e20b1f73ed8f6285fe3268b59
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>