QuickShapes: Handle removals of ShapePaths correctly
Removing paths other than the last could put the the renderers out of
sync, since the path items would change their place in the list. Fix
by ensuring that all items added to the list are marked as dirty. This
works since removals are implemented as list clearing followed by
adding back the unremoved items.

In addition, the curve renderer did not remove the scene graph nodes
when the path list shrinks. Same for the geometry renderer when the
list shrinks to empty. So removed paths would still be visible. Fix by
implementing handling of these cases.

Fixes: QTBUG-133230
Fixes: QTBUG-133231
Pick-to: 6.9 6.8
Change-Id: Ie871e480b99c6b22be9efbceb1e3a4f828f868e7
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>