Doc: Clarify use of QQuickWindow::beforeRendering
The SceneGraph - RHI under QML example gives the impression that
QQuickWindow::beforeRendering() is used with underlays and
QQuickWindow::afterRendering() is used with overlays which is
misleading.

QQuickWindow::beforeRendering() can be used to upload data to buffers in
both instances. Then, based on whether an underlay/overlay is required,
the user can QQuickWindow::beforeRenderPassRecording() or
QQuickWindow::afterRenderPassRecording(), respectively.

This patch updates the documentation to make this distinction clear.

Fixes: QTBUG-130570
Pick-to: 6.6 6.7 6.8 6.9
Change-Id: I3d086b9309d759a5e80f4abbd81d2fc1da55e9f0
Reviewed-by: Michal Klocek <michal.klocek@qt.io>