qmlformat: Show line diffs WRT the expected in writeOut stability checks
The arguments to the diff were inverted meaning that the additions and
removals in the diff were flipped. They should show deviations from the
expected, not the other way around.

```
import QtQml

QtObject /**/ {
    function f() {
    }
}
```

In-place formatting of the above file currently fails the stability
tests because it inserts an extra space. It now shows the correct
direction for the diff:

```
-QtObject /**/  {
-    function f() {
+QtObject /**/   {
+    function f() {
```

Change-Id: I835ffe6c267662b1bb6cc2481f32f55a0b43362f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>