line-by-line format: Fix state transitions for property definitions
The line-by-line formatter would in certain cases considier that a
property definition was completed after having seen only the `property`
token. Fix the state transition to also handle T_RESERVED_WORD tokens.

Picking this back to previous versions is not trivial as this particular
situation seems to break in different ways first.

Fixes: QTBUG-133398
Pick-to: 6.9
Change-Id: I1123fe41e55d2d31c5d245dafe6c8684bca93e39
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>