tst_reformater: Fail test if input or output files are not found
The current setup is very misleading. You could for example write a test
but have a typo in the file path referencing it. The test will still
pass and you would be satisfied in the idea that you cemented the
behavior whereas nothing was actually tested. A warning is not enough
with the amount of output the tests generate.

Fail the test on missing files.

Change-Id: I16913b422ba6ca3b891c38a825beda94d1c0443a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>